Use jsoniter for responding. Use Prometheus codecs for that. Benchmark:

```
goos: linux
goarch: amd64
pkg: github.com/thanos-io/thanos/pkg/api
cpu: Intel(R) Core(TM) i9-10885H CPU @ 2.40GHz
           │     old      │                newv2                 │
           │    sec/op    │    sec/op     vs base                │
Respond-16   1.497m ± 96%   1.208m ± 14%  -19.32% (p=0.000 n=10)

           │       old       │                newv2                 │
           │      B/op       │     B/op      vs base                │
Respond-16   1432.4Ki ± 227%   259.1Ki ± 1%  -81.91% (p=0.000 n=10)

           │     old     │               newv2                │
           │  allocs/op  │ allocs/op   vs base                │
Respond-16   30.000 ± 0%   8.000 ± 0%  -73.33% (p=0.000 n=10)
```

<!--
    Keep PR title verbose enough and add prefix telling
    about what components it touches e.g "query:" or ".*:"
-->

<!--
    Don't forget about CHANGELOG!

    Changelog entry format:
    - [#<PR-id>](<PR-URL>) Thanos <Component> ...

    <PR-id> Id of your pull request.
    <PR-URL> URL of your PR such as https://github.com/thanos-io/thanos/pull/<PR-id>
    <Component> Component affected by your changes such as Query, Store, Receive.
-->

* [ ] I added CHANGELOG entry for this change.
* [ ] Change is not relevant to the end user.

## Changes

<!-- Enumerate changes you made -->

## Verification

<!-- How you tested it? How do you know it works? -->